refactor: AI 평가 점수 float 처리

+++ backend/src/main/java/com/inveskit/backend/dto/AnalysisResponse.java
@@ -15,5 +15,5 @@ public class AnalysisResponse {
     private String signal;
     private String evaluation;
     private String advice;
-    private Map<String, Integer> scores;
+    private Map<String, Double> scores;
 }
